Successful dev run of new LLC4320 benchmarks for Zarr on gcsfs and FUSE.
