+++ benchmarks/IO_numpy.py
